Keep the formation projections beside their consumers
